pyproject.toml
[tool.poetry]
name = "dns-switcher-project"
version = "0.1.0"
description = ""
authors = ["sonya <sonyakarm@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
logging = "^0.4.9.6"
fastapi-jsonapi = "^2.8.0"
flake8 = "^7.1.1"
aiohttp = "^3.11.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
